// src/interfaces/seo.interface.ts
export interface SiteMetadata {
  title: string;
  description: string;
  author: string;
  siteUrl: string;
  ogImage: string;
}
export interface SeoData {
  title: string;
  lang: string;
  description: string;
  noIndex: boolean;
  urlCanonical: string;
  ogImage: string;
  slug: string;
  author: string;
}